* sanxei95 - some refactoring
