+++ src/main/kotlin/com/kos/tasks/runners/CacheGameDataTaskRunner.kt
@@ -13,8 +13,9 @@ import com.kos.tasks.repository.TasksRepository
 import com.kos.views.Game
 import java.time.OffsetDateTime
 
-abstract class CacheGameDataTaskRunner(
-    protected val game: Game,
+class CacheGameDataTaskRunner(
+    private val game: Game,
+    override val type: TaskType,
     private val tasksRepository: TasksRepository,
     private val syncEntitySelector: SyncEntitySelector,
     private val entitySynchronizerProvider: EntitySynchronizerProvider
@@ -30,48 +31,11 @@ abstract class CacheGameDataTaskRunner(
                 right = { it.synchronize(entities) }
             )
 
-        if (errors.isEmpty()) {
-            tasksRepository.updateTask(
-                Task(
-                    id,
-                    type,
-                    TaskStatus(Status.SUCCESSFUL, "entities synced: ${entities.map { it.id }.joinToString(",")}"),
-                    OffsetDateTime.now()
-                )
-            )
-        } else {
-            tasksRepository.updateTask(
-                Task(
-                    id,
-                    type,
-                    TaskStatus(Status.ERROR, errors.joinToString(",\n") { it.toString() }),
-                    OffsetDateTime.now()
-                )
-            )
-        }
+        val taskStatus = if (errors.isEmpty())
+            TaskStatus(Status.SUCCESSFUL, "entities synced: ${entities.map { it.id }.joinToString(",")}")
+        else
+            TaskStatus(Status.ERROR, errors.joinToString(",\n") { it.toString() })
+
+        tasksRepository.updateTask(Task(id, type, taskStatus, OffsetDateTime.now()))
     }
-}
-
-class CacheLolDataTaskRunner(
-    tasksRepository: TasksRepository,
-    syncEntitySelector: SyncEntitySelector,
-    entitySynchronizerProvider: EntitySynchronizerProvider
-) : CacheGameDataTaskRunner(Game.LOL, tasksRepository, syncEntitySelector, entitySynchronizerProvider) {
-    override val type = TaskType.CACHE_LOL_DATA_TASK
-}
-
-class CacheWowDataTaskRunner(
-    tasksRepository: TasksRepository,
-    syncEntitySelector: SyncEntitySelector,
-    entitySynchronizerProvider: EntitySynchronizerProvider
-) : CacheGameDataTaskRunner(Game.WOW, tasksRepository, syncEntitySelector, entitySynchronizerProvider) {
-    override val type = TaskType.CACHE_WOW_DATA_TASK
-}
-
-class CacheWowHcDataTaskRunner(
-    tasksRepository: TasksRepository,
-    syncEntitySelector: SyncEntitySelector,
-    entitySynchronizerProvider: EntitySynchronizerProvider
-) : CacheGameDataTaskRunner(Game.WOW_HC, tasksRepository, syncEntitySelector, entitySynchronizerProvider) {
-    override val type = TaskType.CACHE_WOW_HC_DATA_TASK
 }
